Quantity is updated to integer value in the GRN screens.

+++ app/src/main/java/com/lazymohan/zebraprinter/ZPLPrinterActivity.kt
@@ -67,7 +67,7 @@ class ZPLPrinterActivity : ComponentActivity() {
             val gtinNumber: String? = intent.getStringExtra("gtinNumber")
             val qty: String? = intent.getStringExtra("qty")
             if (!qty.isNullOrEmpty()) {
-                viewModel.updateNoOfCopies(qty)
+                viewModel.updateNoOfCopies(qty.toInt())
             }
             TUITheme {
                 PrinterScreenContent(
